requirements.txt
Pillow>=10.0.0
opencv-python-headless>=4.8.1.0
face-recognition>=1.3.0